Renamed provider_product_stock model and route to stock, added functionality to get the stock from a certain product and redefined get routes to work as intended
