+++ models/Stock.js
@@ -1,7 +1,7 @@
 const DataTypes = require('sequelize');
 
 module.exports = (sequelize) => {
-    sequelize.define('provider_product_stock',{
+    sequelize.define('stock',{
         quantity: {
             type: DataTypes.INTEGER,
             allowNull: false,
