+++ package.json
@@ -17,6 +17,7 @@
     "express": "^4.17.3",
     "jsonwebtoken": "^8.5.1",
     "nodemailer": "^6.7.3",
+    "nodemon": "^2.0.15",
     "pg": "^8.7.3",
     "pg-hstore": "^2.3.4",
     "sequelize": "^6.18.0"
