+++ routes/stock.js
